requirements.txt
numpy
scipy
libigl
sklearn
tqdm
trimesh